Move dependency-injection-related files into a di subdirectory
