+++ package.json
@@ -62,7 +62,7 @@
     "start:web-client": "cd packages/web-client && ember serve",
     "test": "npm-run-all --aggregate-output --continue-on-error --parallel test:*",
     "test:did-resolver": "PACKAGE=did-resolver mocha ./packages/test-support/bin/run.js --timeout 60000",
-    "test:hub": "PACKAGE=hub mocha ./packages/test-support/bin/run.js --timeout 60000",
+    "test:hub": "cd packages/hub && PACKAGE=hub NODE_ENV=test mocha ../../packages/test-support/bin/run.js --timeout 60000",
     "test:web-client": "cd packages/web-client && ember test",
     "test:web-client:serve": "cd packages/web-client && ember test --serve --no-launch"
   },
