+++ packages/hub/jsonapi-middleware.ts
@@ -8,7 +8,7 @@ import { Memoize } from 'typescript-memoize';
 import { CardstackError } from './error';
 import { SessionContext } from './authentication-middleware';
 import SessionRoute from './routes/session';
-import { inject } from './dependency-injection';
+import { inject } from './di/dependency-injection';
 
 const API_PREFIX = '/api';
 const apiPrefixPattern = new RegExp(`^${API_PREFIX}/(.*)`);
@@ -63,7 +63,7 @@ export default class JSONAPIMiddleware {
   }
 }
 
-declare module '@cardstack/hub/dependency-injection' {
+declare module '@cardstack/hub/di/dependency-injection' {
   interface KnownServices {
     'jsonapi-middleware': JSONAPIMiddleware;
   }
