+++ packages/hub/routes/session.ts
@@ -1,5 +1,5 @@
 import Koa from 'koa';
-import { inject } from '../dependency-injection';
+import { inject } from '../di/dependency-injection';
 import { AuthenticationUtils } from '../utils/authentication';
 import NonceTracker, { MAX_NONCE_AGE_NS } from '../nonce-tracker';
 import { recoverTypedSignature } from 'eth-sig-util';
@@ -96,7 +96,7 @@ export default class SessionRoute {
   }
 }
 
-declare module '@cardstack/hub/dependency-injection' {
+declare module '@cardstack/hub/di/dependency-injection' {
   interface KnownServices {
     'session-route': SessionRoute;
   }
