+++ packages/hub/utils/authentication.ts
@@ -1,4 +1,4 @@
-import { inject } from '../dependency-injection';
+import { inject } from '../di/dependency-injection';
 import { createCipheriv, createDecipheriv, createHmac, randomBytes } from 'crypto';
 import { Clock } from './clock';
 import queryString from 'query-string';
@@ -80,7 +80,7 @@ export class AuthenticationUtils {
   }
 }
 
-declare module '@cardstack/hub/dependency-injection' {
+declare module '@cardstack/hub/di/dependency-injection' {
   interface KnownServices {
     'authentication-utils': AuthenticationUtils;
   }
